% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\name{convertFromTo}
\alias{convertFromTo}
\title{Convert molecular identifier to target identifiers}
\usage{
convertFromTo(gng, queryId, targetDb = c("GeneSymbolDB"), sourceDb = NA,
  longFormat = TRUE, chromosomal = TRUE)
}
\arguments{
\item{gng}{A GeneNameGenieR object (e.g. \code{GeneNameGenieR()})}

\item{queryId}{A character vector of molecular identifiers that needs to be converted}

\item{targetDb}{A character vecotr of target databases (default equals to 'EntrezGene' and 'GeneSymbolDb' for
the Entrez gene ID and official gene symbol respectively)}

\item{sourceDb}{String containing the origin database of the input ID(s) (optional).}

\item{longFormat}{Boolean parameter (default TRUE). Indicates if returned dataframe is in the form of a longFormat}

\item{chromosomal}{Boolean parameter (default TRUE). Whether the search and output shall include features located on
chromosomes other than the standard 23 {1,...21, X, Y}. Ensembl also includes features that are located on so called
alternative human chromosomes, such as CHR_HSCHR5_6_CTG1. In the case where \preformatted{chromosomal} is \preformatted{FALSE}, not even input
identifiers that are associated with this location will be found.}
}
\value{
A data.frame object with the columns InputId, InputSourceDb and columns containing TargetDb and TargetId values.
}
\description{
Converts molecular input identifier to identifier from target database. Thereby, the source database of the
input identifier does not have to be specified as the database detects the corresponding database automatically.
}
\examples{
\dontrun{
  gng = GeneNameGenieR();
  convertFromTo(gng, c("BCL2", "AMPK"), "EntrezGene");
  InputId        InputSourceDb  TargetDb TargetId
1    BCL2 Official Gene Symbol NCBI gene      596
2    AMPK    Gene Symbol Alias NCBI gene     5563
}

}
\seealso{
\code{\link{query}}
}
