% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\docType{methods}
\name{GeneNameGenieR,ANY-method}
\alias{GeneNameGenieR,ANY-method}
\title{GeneNameGenieR constructor}
\usage{
\S4method{GeneNameGenieR}{ANY}(...)
}
\arguments{
\item{url}{: Graph database URL, such as "http://localhost:7474/db/data/"}
}
\value{
an object of class 'GeneNameGenieR'
}
\description{
This function returns an instance of a GeneNameGenieR class.
}
\details{
This function initializes an object of the class GeneNameGenieR It is a
wrapper for \code{new()}.
}
\examples{
\dontrun{
gng = GeneNameGenieR() # Instance of class 'GeneNameGenieR'
}
}
\seealso{
\code{\link{new}}
}
\author{
Stefan J. Haunsberger
}
