% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\docType{data}
\name{GeneNameGenieR}
\alias{GeneNameGenieR}
\alias{DEFAULT_URL}
\title{GeneNameGenieR: Lightning fast graph based gene name converter}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_URL
}
\description{
For bug reports, please
\href{https://github.com/xxx/xxx/issues}{create a new GitHub issue}.
For support, please \href{http://stackoverflow.com/questions/tagged/xxx}{create a question on StackOverflow}.
}
\author{
Stefan J. Haunsberger
}
\keyword{datasets}
