% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getCurrentMirbaseVersion}
\alias{getCurrentMirbaseVersion}
\title{Get highest supported miRBase release version information}
\usage{
getCurrentMirbaseVersion()
}
\examples{
\dontrun{
  getCurrentMirbaseVersion();
}
}
