% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\name{getOfficialGeneSymbol}
\alias{getOfficialGeneSymbol}
\title{Convert input id to its corresponding official gene symbol}
\usage{
getOfficialGeneSymbol(gng, queryId, sourceDb = NA_character_,
  chromosomal = TRUE)
}
\arguments{
\item{gng}{A GeneNameGenieR object (e.g. \code{GeneNameGenieR()})}

\item{queryId}{A character vector of molecular identifiers that needs to be converted}

\item{sourceDb}{String containing the origin database of the input ID(s) (optional).}

\item{chromosomal}{Boolean parameter (default TRUE). Whether the search and output shall include features located on
chromosomes other than the standard 23 {1,...21, X, Y}. Ensembl also includes features that are located on so called
alternative human chromosomes, such as CHR_HSCHR5_6_CTG1. In the case where \preformatted{chromosomal} is \preformatted{FALSE}, not even input
identifiers that are associated with this location will be found.}

\item{targetDb}{A character vector of target databases (default equals to 'EntrezGene' and 'GeneSymbolDb' for
the Entrez gene ID and official gene symbol respectively)}
}
\description{
Compute the official gene symbol for a single or multiple `queryId`(s).
If `sourceDb` is missing, the database will automatically try to detect the queryId's
  source database. Per default `chromosomal` is set to `true`.
}
