% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getValidDatabases}
\alias{getValidDatabases}
\title{Show all supported DBs}
\usage{
getValidDatabases()
}
\value{
A \preformatted{data.frame} object with the columns DatabaseDisplayName and DatabaseID.
}
\description{
Returns a table containing all supported DBs with DB name and ID. The ID can
used for example as targetDbs parameter in the \code{query} and \code{convertFromTo} function.
}
\examples{
\dontrun{
  getValidDatabases();
}

}
\seealso{
\code{\link{query}} and \code{\link{convertFromTo}}
}
