% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getValidMirnaMetadataValues}
\alias{getValidMirnaMetadataValues}
\title{Available attributes}
\usage{
getValidMirnaMetadataValues()
}
\value{
A \preformatted{data.frame} object with the columns DatabaseDisplayName and DatabaseID.
}
\description{
Returns a table containing all supported attributes that can serve as
'params' input parameter in the \code{\link{query}} function.
}
\examples{
\dontrun{
  getValidMirnaMetadataValues();
}
}
\seealso{
\code{\link{query}}
}
