% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\name{query}
\alias{query}
\title{Translate identifiers and receive attributes for Ensembl genes, transcripts and proteins}
\usage{
query(gng, queryId, targetDb = c("GeneSymbolDB"), params = NA,
  sourceDb = NA, longFormat = TRUE, chromosomal = TRUE)
}
\arguments{
\item{gng}{An instance of the GeneNameGenieR object (e.g. \code{gng = GeneNameGenieR()})}

\item{queryId}{A \preformatted{character} vector of molecular identifiers that needs to be converted}

\item{params}{String or vector containing retrievable attributes, such as 'engs:region' for the Ensembl gene region (\code{\link{getValidAttributes}}).}

\item{sourceDb}{String containing the origin database of the input ID(s) (optional).}

\item{targetDbs}{A \preformatted{character} vector of target databases (default equals to 'EntrezGene' and 'GeneSymbolDb' for
the Entrez gene ID and official gene symbol respectively)}
}
\value{
A \preformatted{data.frame} object with the columns InputId, InputSourceDb, TargetDb and TargetId.
}
\description{
Similar to the \code{\link{convertFromTo}} function, the \code{\link{query}} function converts any matchable
input identifier to the desired target database identifier. Moreover, with the \preformatted{params} parameter
\code{\link{query}} returns attributes for Ensembl genes, transcripts and proteins, such as location and biotype.
}
\examples{

\dontrun{
query(
   GeneNameGenieR(),
   c('AMPK'))
}
\dontrun{
query(
   GeneNameGenieR(),
   c('AMPK', 'Bcl-2', '596', 'NM_000657'),
   targetDb = c('EntrezGene', 'ArrayExpress', 'RefSeq_mRNA'),
   params = c('ensg:region', 'ensg:regionStart', 'ensg:regionEnd'),
   longFormat = FALSE)

     InputId     n
       <chr> <int>
   1     596     2
     InputId     InputSourceDb  GeneEnd GeneStart   EnsemblGeneId GeneRegion NCBI.gene RefSeq.mRNA
1        596          WikiGene 63320128  63123346 ENSG00000171791         18       596   NM_000657
2        596          WikiGene 63320128  63123346 ENSG00000171791         18       596   NM_000633
3        596          WikiGene 63320128  63123346 ENSG00000171791         18       596        <NA>
4        596         NCBI gene 63320128  63123346 ENSG00000171791         18       596   NM_000657
5        596         NCBI gene 63320128  63123346 ENSG00000171791         18       596   NM_000633
6        596         NCBI gene 63320128  63123346 ENSG00000171791         18       596        <NA>
7       AMPK Gene Symbol Alias 56715335  56645322 ENSG00000162409          1      5563        <NA>
8       AMPK Gene Symbol Alias 56715335  56645322 ENSG00000162409          1      5563   NM_006252
9      Bcl-2 Gene Symbol Alias 63320128  63123346 ENSG00000171791         18       596   NM_000657
10     Bcl-2 Gene Symbol Alias 63320128  63123346 ENSG00000171791         18       596   NM_000633
11     Bcl-2 Gene Symbol Alias 63320128  63123346 ENSG00000171791         18       596        <NA>
12 NM_000657       RefSeq mRNA 63320128  63123346 ENSG00000171791         18       596   NM_000657
Warning message:
In postCheckInput(x) :
Some input identifiers match to more than one input source database
}
}
\seealso{
\code{\link{convertFromTo}}
}
