% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpea.R
\name{rpea}
\alias{rpea}
\title{Perform Reactome Pathway Enrichment Analysis (RPEA)}
\usage{
rpea(queryIds, sourceDb = "GeneSymbolDB", pValueCutoff = 0.05,
  minNAnnotatedGenes = 10, maxNAnnotatedGenes = 500,
  pAdjustmentMethod = "BF", qValueCutoff = 0.2)
}
\arguments{
\item{queryIds}{A character vector of molecular identifiers.}

\item{sourceDb}{String containing the origin database of the input ID(s) (optional).}

\item{pValueCutoff}{Floating point value (default 0.05) for p-value threshold.}

\item{minNAnnotatedGenes}{Integer value (default 10) denoting the minimum number of annotated genes a Reactome
pathway needs to be included in the analysis.}

\item{maxNAnnotatedGenes}{Integer value (default 500) denoting the maximum number of annotated genes a Reactome
pathway needs to be included in the analysis.}

\item{pAdjustmentMethod}{String (default 'BF' - Bonferroni) containing the p-value correction for multiple testing
(other options: 'BH' - Benjamini and Hochberg)}
}
\value{
A data.frame object with the columns InputId, InputSourceDb and columns containing TargetDb and TargetId values.
}
\description{
Run the Cypher stored procedure `rcsi.enrichment.reactomePathwayEnrichment`.
}
\examples{
\dontrun{
  GeneNameGenieR::rpea(c('PRKAA1', 'PRKAB2', 'PRKAG2', 'STK11'))
  InputId        InputSourceDb  TargetDb TargetId
1    BCL2 Official Gene Symbol NCBI gene      596
2    AMPK    Gene Symbol Alias NCBI gene     5563
}


}
