% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\name{setNeo4jConnection}
\alias{setNeo4jConnection}
\title{Set GeneNameGenie-Neo4j connection}
\usage{
setNeo4jConnection(host = DEFAULT_HOST, port = DEFAULT_PORT,
  path = DEFAULT_PATH)
}
\arguments{
\item{host}{A string value of the host, such as 'localhost' or an IP address}

\item{port}{An integer value for the port, such as 7474 (default)}

\item{path}{String value containing the path to the database files (defaults to `db/data/`). The value must end on a `/` (slash).}
}
\description{
The connection defaults to 'localhost' if not specified differently.
}
