% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\docType{data}
\name{GeneNameGenieR}
\alias{GeneNameGenieR}
\alias{DEFAULT_URL}
\alias{GeneNameGenieR}
\title{GeneNameGenieR: Lightning fast graph based gene name converter}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_URL

GeneNameGenieR(url = NA_character_)
}
\arguments{
\item{url}{: Graph database URL, such as "http://localhost:7474/db/data/"}
}
\description{
For bug reports, please
\href{https://github.com/StefanHaunsberger/GeneNameGenieR/issues}{create a new GitHub issue}.
}
\author{
Stefan J. Haunsberger
}
\keyword{datasets}
