% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\docType{package}
\name{GeneNameGenieR}
\alias{GeneNameGenieR}
\alias{GeneNameGenieR-package}
\alias{GeneNameGenieR}
\title{GeneNameGenieR: Lightning fast graph based gene name converter}
\usage{
GeneNameGenieR(...)
}
\description{


This function returns back an instance of a
\emph{MiRNANAmeConverter} object.
}
\section{Slots}{

\describe{
\item{\code{url}}{Database connection}

\item{\code{graph}}{Valid/Supported miRBase versions}

\item{\code{toAddDbCols}}{Number of different organisms supported}
}}

\author{
Stefan J. Haunsberger

Stefan J. Haunsberger

For bug reports, please
\href{https://github.com/StefanHaunsberger/GeneNameGenieR/issues}{create a new GitHub issue}.
}
