% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameGenieR.R
\docType{package}
\name{GeneNameGenieR}
\alias{GeneNameGenieR}
\alias{GeneNameGenieR-package}
\title{GeneNameGenieR: Lightning fast graph based gene name converter}
\description{

}
\author{
Stefan J. Haunsberger
}
