% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getValidMirnaMetadataValues}
\alias{getValidMirnaMetadataValues}
\title{Available attributes}
\usage{
getValidMirnaMetadataValues(gng)
}
\arguments{
\item{gng}{An instance of the \code{GeneNameGenieR} object (e.g. \preformatted{gng = new("GeneNameGenieR")})}
}
\value{
A \preformatted{data.frame} object with the columns DatabaseDisplayName and DatabaseID.
}
\description{
Returns a table containing all supported attributes that can serve as
'params' input parameter in the \code{\link{query}} function.
}
\examples{
\dontrun{
  gng = GeneNameGenieR();
  getValidMirnaMetadataValues(gng);
}
}
\seealso{
\code{\link{query}}
}
